Dedup initialState.
