+++ packages/server/src/time-control/time-handler-sequential.ts
@@ -6,11 +6,11 @@ import {
   IPerPlayerTimeControlBase,
   ITimeControlBase,
   TimeControlType,
-  makeGameObject,
 } from "@ogfcommunity/variants-shared";
 import { ITimeoutService } from "./timeout";
 import { ITimeHandler } from "./time-control";
 import { IClock } from "./clock";
+import { initialState } from "./time-handler-utils";
 
 export class TimeHandlerSequentialMoves implements ITimeHandler {
   private _timeoutService: ITimeoutService;
@@ -25,37 +25,7 @@ export class TimeHandlerSequentialMoves implements ITimeHandler {
     variant: string,
     config: IConfigWithTimeControl,
   ): ITimeControlBase {
-    const numPlayers = makeGameObject(variant, config).numPlayers();
-
-    const timeControl: ITimeControlBase = {
-      moveTimestamps: [],
-      forPlayer: {},
-    };
-
-    for (let i = 0; i < numPlayers; i++) {
-      timeControl.forPlayer[i] = {
-        remainingTimeMS: config.time_control.mainTimeMS,
-        onThePlaySince: null,
-      };
-    }
-
-    switch (config.time_control.type) {
-      case TimeControlType.Absolute: {
-        // nothing to do
-        break;
-      }
-
-      case TimeControlType.Fischer: {
-        // nothing to do
-        break;
-      }
-
-      default: {
-        console.error("received config with invalid time control type");
-      }
-    }
-
-    return timeControl;
+    return initialState(variant, config);
   }
 
   handleMove(
